\docType{data}
\name{rain_max_world}
\alias{rain_max_world}
\title{Record rainfall totals from around the world}
\format{Data frame with 8 columns and 47 rows}
\description{
  World record point precipitation measurements from NOAA's
  National Weather Service, Hydrometeorological Design
  Studies Center. NOAA indicates that some records have not
  been verified.
}
\details{
  Variables:

  \itemize{ \item Duration - duration of the rainfall event
  \item Amount_in - total rainfall amount in inches \item
  Amount_mm - total rainfall amount in millimeters \item
  Location - region and country names \item Lat - latitude
  in decimal degrees \item Lon - longitude in decimal
  degrees \item Start_Date - starting date of the rainfall
  event in yyyy-mm-dd format \item Estimate - Yes if the
  record has not been verified, blank otherwise }
}
\author{
  Gopi Goteti
}
\references{
  World record point precipitation measurements,
  \url{http://www.nws.noaa.gov/oh/hdsc/record_precip/record_precip_world.html},
  extracted May 18 2014.
}
\keyword{datasets}

