\docType{package}
\name{rainfreq}
\alias{rainfreq}
\alias{rainfreq-package}
\title{Rainfall frequency estimates for the USA from the NOAA National Weather
Service (NWS) division Hydrometeorological Design Studies Center (HDSC).}
\description{
  Rainfall frequency estimates for the USA from the NOAA
  National Weather Service (NWS) division
  Hydrometeorological Design Studies Center (HDSC).
}
\details{
  Data from NOAA NWS is available in various formats.
  \pkg{rainfreq} provides functionality to access the GIS
  format files provided by NWS' PF Data Server.
  \pkg{rainfreq} regional selection criterion is currently
  limited to the 50 states (plus DC). \pkg{rainfreq} also
  comes with datasets on record point rainfall measurements
  provided by NWS-HDSC
  \url{http://www.nws.noaa.gov/oh/hdsc/record_precip/record_precip.html}
}
\author{
  Gopi Goteti
}
\references{
  NOAA NWS HDSC
  \url{http://www.nws.noaa.gov/oh/hdsc/index.html}. Data in
  GIS format is available from the PF Data Server
  \url{http://hdsc.nws.noaa.gov/hdsc/pfds/pfds_gis.html}.
}

